feat: remove dependency between auth and keys package
We define the interface for copier renewer where its being used in the auth package so that any structure can satisfy its needs without being dependent on the keys package.

The keys package instead returns a struct from its broker function with an additional method for deferring a close on the broker in favour of the extra return value.

As a result of this the mock had to move out to a separate package, which overall cleaned up the rest of the keys broker code.