core/env: add alt func to overload/overwrite existing env vars.
Currently, the infra/.env file take precedence over infra/local.env, without
the option to reverse them. This change provides an alternative function
that will override env variable with the definition given later. This has
the effect of allowing the local.env, OR files listed by the caller, to
overload.

The current implementation suggests that individual files will be skipped
if missing, but that is not the case. godotenv's Load/Unload calls will
return an error on the first missing file encountered, and will thus never
attempt later files. This change instead ranges over the files locally,
passing one to godotenv at a time. This provides the expected behaviour,
skipping missing files and attempting the next.