automatically handle CORS preflight requests, globally
This change makes is so that the server, when run, will automatically
accept and handle all CORS requests. The added CORS handler responds to
CORS preflight requests with the needed headers updated to allow CORS
requests. When running the server the CORS handler is added such that
all requests with the method OPTIONS will be treated as CORS preflight
requests.

The pkg user can also modify the CORS origin/headers/methods directly
after creating the Server with NewServer().