workers/keybroker: synchronise access to RSA broker's keys
The RSA public/private key brokers are racing on the key field.
Copying, checking, and resetting this field can all be done
concurrently, and there is currently no sychronisation. This seems to me
like an appropriate use of a mutex, as we are simply writing/reading the
state of a field, not passing ownership or communicating results.

Fixes #97