workers/keybroker: sync access to broker's running state
The RSA key broker's check whether their underlying Broker is running or
not by reading the Broker's field `running`. Because the RSA broker and
the Broker are running concurrently, this check may conflict with the
Broker's write to that same field when updating it's running state.

This change adds a mutex to synchronise access to that field. I've used
a smalle helper function to allow the parent RSA broker to check the
running status of it's underlying Broker. Because the check is in the if
statement, the naked use of the mutex looks odd (you lock before the if
statement, but have to unlock twice, one in each branch of the if
statement). Also, the RSA Broker is reaching in and checking the state
of an underlying Broker, so I thought it was nicer to have that
underlying Broker handle it's own mutex.

Updates #97